web-sockets/pyproject.toml
[tool.poetry]
name = "webcourse-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.0"
uvicorn = "^0.22.0"
pydantic = "2.3.0"
requests = "^2.31.0"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
asyncpg = "^0.28.0"
SQLAlchemy = "^2.0.17"
gunicorn = "^20.1.0"
async_fastapi_jwt_auth="0.6.4"
PyJWT = "^2.8.0"
prometheus-fastapi-instrumentator = "^6.0.0"
starlette-exporter = "^0.16.0"
sentry-sdk = { extras = ["fastapi"], version = "^1.26.0" }
aioboto3 = "^11.2.0"
bcrypt="^4.1.2"
pydantic-settings = "^2.0.1"
python-multipart = "^0.0.9"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.2"
ipython = "^8.14.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.cruft]
skip = [".env"]

[tool.pytest.ini_options]
asyncio_mode = 'auto'
